Fixed union find bugs and consolidated code
